Add req fields + auto update minor version of make-agent (#28)
When deploying an agent using the current setup, the process fails if a
repository or email is not provided. This PR adds clarifying comments to
help guide users through these requirements. Additionally, the
deployment file referenced an ACCOUNT_ID that was missing from the
.env.example, which could easily cause confusion. That has now been
addressed to improve developer onboarding and reduce setup errors.

---------

Co-authored-by: Nate Geier <nate@mintbase.xyz>